[Refactor] #4166 colored-char.h をクラス名に合わせてdisplay-symbol.h に変えた
