+++ Hengband/Hengband/Hengband.vcxproj.filters
@@ -3716,7 +3716,7 @@
     <ClInclude Include="..\..\src\cmd-io\cmd-lore.h">
       <Filter>cmd-io</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\view\colored-char.h">
+    <ClInclude Include="..\..\src\view\display-symbol.h">
       <Filter>view</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\view\display-monster-status.h">
