+++ src/system/baseitem-info.h
@@ -5,7 +5,7 @@
 #include "object/tval-types.h"
 #include "system/angband.h"
 #include "util/flag-group.h"
-#include "view/colored-char.h"
+#include "view/display-symbol.h"
 #include <array>
 #include <optional>
 #include <string>
