+++ src/view/display-symbol.h
